REFACTOR: Change HttpStage open to take a return status code.
If HttpStage.open gets an error, its close() callback should not be invoked. To handle this, open must return a status code.
If the handler open() calls httpError, the passHandler needs to be used instead thereafter.
Also need more robust pendingFinalize if httpError is called before the pipeline is fully constructed.